Some more structure to the dashboard via redeligating components and styles to their own directories
