+++ client/src/components/UserProgressCard.js
@@ -1,0 +1,47 @@
+import React from 'react';
+import { useTheme } from '@material-ui/core/styles';
+import Card from '@material-ui/core/Card';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import IconButton from '@material-ui/core/IconButton';
+import Typography from '@material-ui/core/Typography';
+import SkipPreviousIcon from '@material-ui/icons/SkipPrevious';
+import PlayArrowIcon from '@material-ui/icons/PlayArrow';
+import SkipNextIcon from '@material-ui/icons/SkipNext';
+import { useStyles } from '../styles/UserProgressCardStyles'
+
+export default function UserProgressCard() {
+    const classes = useStyles();
+    const theme = useTheme();
+
+    return (
+        <Card className={classes.root}>
+            <div className={classes.details}>
+                <CardContent className={classes.content}>
+                    <Typography component="h5" variant="h5">
+                        Live From Space
+          </Typography>
+                    <Typography variant="subtitle1" color="textSecondary">
+                        Mac Miller
+          </Typography>
+                </CardContent>
+                <div className={classes.controls}>
+                    <IconButton aria-label="previous">
+                        {theme.direction === 'rtl' ? <SkipNextIcon /> : <SkipPreviousIcon />}
+                    </IconButton>
+                    <IconButton aria-label="play/pause">
+                        <PlayArrowIcon className={classes.playIcon} />
+                    </IconButton>
+                    <IconButton aria-label="next">
+                        {theme.direction === 'rtl' ? <SkipPreviousIcon /> : <SkipNextIcon />}
+                    </IconButton>
+                </div>
+            </div>
+            <CardMedia
+                className={classes.cover}
+                image="/static/images/cards/live-from-space.jpg"
+                title="Live from space album cover"
+            />
+        </Card>
+    );
+}
