--- client/src/components/copyright/copyright.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-import Typography from '@material-ui/core/Typography';
-import MuiLink from '@material-ui/core/Link';
-
-const Copyright = () => {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {'Copyright © '}
-      <MuiLink color="inherit" href="https://material-ui.com/">
-        Your Website
-      </MuiLink>{' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-};
-
-export default Copyright;
