--- client/src/components/copyright/index.js
@@ -1,1 +1,0 @@
-export { default } from './copyright';
