--- client/src/components/flex/index.js
@@ -1,1 +1,0 @@
-export { default } from './flex';
