--- client/src/components/link/index.js
@@ -1,1 +1,0 @@
-export { default } from './link';
