--- client/src/components/pro-tip/index.js
@@ -1,1 +1,0 @@
-export { default } from './pro-tip';
