+++ client/src/styles/UserProgressCardStyles.js
@@ -1,0 +1,27 @@
+import { makeStyles, useTheme } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles((theme) => ({
+    root: {
+        display: 'flex',
+    },
+    details: {
+        display: 'flex',
+        flexDirection: 'column',
+    },
+    content: {
+        flex: '1 0 auto',
+    },
+    cover: {
+        width: 151,
+    },
+    controls: {
+        display: 'flex',
+        alignItems: 'center',
+        paddingLeft: theme.spacing(1),
+        paddingBottom: theme.spacing(1),
+    },
+    playIcon: {
+        height: 38,
+        width: 38,
+    },
+}));
